[#165] PromiseKit 도입 후 LoginViewController의 비즈니스 로직 분리
